// src/app/models/AllCustomer.ts
export class AllCustomer{
    userId?: number;
    activeStatus?: number;
    fullName?: string;
    email?: string;
    uniqueServiceNumber?: string;
    boardName?: string;
    stateId?: number;
    boardId?: number;
    stateName?: string;
    localityName?: string;
    localityId?: number;
    sellOrderCount?: number;
    contractCounts?: number;
}